+++ src/Symfony/Bundle/SecurityBundle/Tests/DependencyInjection/Fixtures/php/bcrypt_hasher.php
@@ -1,0 +1,12 @@
+<?php
+
+$this->load('container1.php');
+
+$container->loadFromExtension('security', [
+    'password_hashers' => [
+        'JMS\FooBundle\Entity\User7' => [
+            'algorithm' => 'bcrypt',
+            'cost' => 15,
+        ],
+    ],
+]);