+++ src/Symfony/Component/Security/Core/Tests/Authentication/Provider/DaoAuthenticationProviderTest.php
@@ -15,17 +15,17 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\Security\Core\Authentication\Provider\DaoAuthenticationProvider;
 use Symfony\Component\Security\Core\Authentication\Token\UsernamePasswordToken;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
-use Symfony\Component\Security\Core\Encoder\PasswordEncoderInterface;
-use Symfony\Component\Security\Core\Encoder\PlaintextPasswordEncoder;
 use Symfony\Component\Security\Core\Exception\AuthenticationServiceException;
 use Symfony\Component\Security\Core\Exception\BadCredentialsException;
 use Symfony\Component\Security\Core\Exception\UsernameNotFoundException;
-use Symfony\Component\Security\Core\Tests\Encoder\TestPasswordEncoderInterface;
 use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 use Symfony\Component\Security\Core\User\User;
 use Symfony\Component\Security\Core\User\UserCheckerInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 use Symfony\Component\Security\Core\User\UserProviderInterface;
+use Symfony\Component\PasswordHasher\Hasher\PasswordHasherFactoryInterface;
+use Symfony\Component\PasswordHasher\Hasher\PlaintextPasswordHasher;
+use Symfony\Component\PasswordHasher\PasswordHasherInterface;
 
 class DaoAuthenticationProviderTest extends TestCase
 {
@@ -39,7 +39,10 @@ class DaoAuthenticationProviderTest extends TestCase
         $method->invoke($provider, 'fabien', $this->getSupportedToken());
     }
 
-    public function testRetrieveUserWhenUsernameIsNotFound()
+    /**
+     * @group legacy
+     */
+    public function testRetrieveUserWhenUsernameIsNotFoundWithLegacyEncoderFactory()
     {
         $this->expectException(UsernameNotFoundException::class);
         $userProvider = $this->createMock(UserProviderInterface::class);
@@ -55,6 +58,22 @@ class DaoAuthenticationProviderTest extends TestCase
         $method->invoke($provider, 'fabien', $this->getSupportedToken());
     }
 
+    public function testRetrieveUserWhenUsernameIsNotFound()
+    {
+        $this->expectException(UsernameNotFoundException::class);
+        $userProvider = $this->createMock(UserProviderInterface::class);
+        $userProvider->expects($this->once())
+            ->method('loadUserByUsername')
+            ->willThrowException(new UsernameNotFoundException())
+        ;
+
+        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(PasswordHasherFactoryInterface::class));
+        $method = new \ReflectionMethod($provider, 'retrieveUser');
+        $method->setAccessible(true);
+
+        $method->invoke($provider, 'fabien', $this->getSupportedToken());
+    }
+
     public function testRetrieveUserWhenAnExceptionOccurs()
     {
         $this->expectException(AuthenticationServiceException::class);
@@ -64,7 +83,7 @@ class DaoAuthenticationProviderTest extends TestCase
                      ->willThrowException(new \RuntimeException())
         ;
 
-        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(EncoderFactoryInterface::class));
+        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(PasswordHasherFactoryInterface::class));
         $method = new \ReflectionMethod($provider, 'retrieveUser');
         $method->setAccessible(true);
 
@@ -85,7 +104,7 @@ class DaoAuthenticationProviderTest extends TestCase
               ->willReturn($user)
         ;
 
-        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(EncoderFactoryInterface::class));
+        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(PasswordHasherFactoryInterface::class));
         $reflection = new \ReflectionMethod($provider, 'retrieveUser');
         $reflection->setAccessible(true);
         $result = $reflection->invoke($provider, 'someUser', $token);
@@ -103,7 +122,7 @@ class DaoAuthenticationProviderTest extends TestCase
                      ->willReturn($user)
         ;
 
-        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(EncoderFactoryInterface::class));
+        $provider = new DaoAuthenticationProvider($userProvider, $this->createMock(UserCheckerInterface::class), 'key', $this->createMock(PasswordHasherFactoryInterface::class));
         $method = new \ReflectionMethod($provider, 'retrieveUser');
         $method->setAccessible(true);
 
@@ -113,13 +132,13 @@ class DaoAuthenticationProviderTest extends TestCase
     public function testCheckAuthenticationWhenCredentialsAreEmpty()
     {
         $this->expectException(BadCredentialsException::class);
-        $encoder = $this->createMock(PasswordEncoderInterface::class);
-        $encoder
+        $hasher = $this->getMockBuilder(PasswordHasherInterface::class)->getMock();
+        $hasher
             ->expects($this->never())
-            ->method('isPasswordValid')
+            ->method('verify')
         ;
 
-        $provider = $this->getProvider(null, null, $encoder);
+        $provider = $this->getProvider(null, null, $hasher);
         $method = new \ReflectionMethod($provider, 'checkAuthentication');
         $method->setAccessible(true);
 
@@ -135,14 +154,14 @@ class DaoAuthenticationProviderTest extends TestCase
 
     public function testCheckAuthenticationWhenCredentialsAre0()
     {
-        $encoder = $this->createMock(PasswordEncoderInterface::class);
-        $encoder
+        $hasher = $this->createMock(PasswordHasherInterface::class);
+        $hasher
             ->expects($this->once())
-            ->method('isPasswordValid')
+            ->method('verify')
             ->willReturn(true)
         ;
 
-        $provider = $this->getProvider(null, null, $encoder);
+        $provider = $this->getProvider(null, null, $hasher);
         $method = new \ReflectionMethod($provider, 'checkAuthentication');
         $method->setAccessible(true);
 
@@ -163,13 +182,13 @@ class DaoAuthenticationProviderTest extends TestCase
     public function testCheckAuthenticationWhenCredentialsAreNotValid()
     {
         $this->expectException(BadCredentialsException::class);
-        $encoder = $this->createMock(PasswordEncoderInterface::class);
-        $encoder->expects($this->once())
-                ->method('isPasswordValid')
+        $hasher = $this->createMock(PasswordHasherInterface::class);
+        $hasher->expects($this->once())
+                ->method('verify')
                 ->willReturn(false)
         ;
 
-        $provider = $this->getProvider(null, null, $encoder);
+        $provider = $this->getProvider(null, null, $hasher);
         $method = new \ReflectionMethod($provider, 'checkAuthentication');
         $method->setAccessible(true);
 
@@ -235,13 +254,13 @@ class DaoAuthenticationProviderTest extends TestCase
 
     public function testCheckAuthentication()
     {
-        $encoder = $this->createMock(PasswordEncoderInterface::class);
-        $encoder->expects($this->once())
-                ->method('isPasswordValid')
+        $hasher = $this->createMock(PasswordHasherInterface::class);
+        $hasher->expects($this->once())
+                ->method('verify')
                 ->willReturn(true)
         ;
 
-        $provider = $this->getProvider(null, null, $encoder);
+        $provider = $this->getProvider(null, null, $hasher);
         $method = new \ReflectionMethod($provider, 'checkAuthentication');
         $method->setAccessible(true);
 
@@ -258,21 +277,21 @@ class DaoAuthenticationProviderTest extends TestCase
     {
         $user = new User('user', 'pwd');
 
-        $encoder = $this->createMock(TestPasswordEncoderInterface::class);
-        $encoder->expects($this->once())
-                ->method('isPasswordValid')
+        $hasher = $this->createMock(PasswordHasherInterface::class);
+        $hasher->expects($this->once())
+                ->method('verify')
                 ->willReturn(true)
         ;
-        $encoder->expects($this->once())
-                ->method('encodePassword')
+        $hasher->expects($this->once())
+                ->method('hash')
                 ->willReturn('foobar')
         ;
-        $encoder->expects($this->once())
+        $hasher->expects($this->once())
                 ->method('needsRehash')
                 ->willReturn(true)
         ;
 
-        $provider = $this->getProvider(null, null, $encoder);
+        $provider = $this->getProvider(null, null, $hasher);
 
         $userProvider = ((array) $provider)[sprintf("\0%s\0userProvider", DaoAuthenticationProvider::class)];
         $userProvider->expects($this->once())
@@ -304,7 +323,7 @@ class DaoAuthenticationProviderTest extends TestCase
         return $mock;
     }
 
-    protected function getProvider($user = null, $userChecker = null, $passwordEncoder = null)
+    protected function getProvider($user = null, $userChecker = null, $passwordHasher = null)
     {
         $userProvider = $this->createMock(PasswordUpgraderProvider::class);
         if (null !== $user) {
@@ -318,18 +337,18 @@ class DaoAuthenticationProviderTest extends TestCase
             $userChecker = $this->createMock(UserCheckerInterface::class);
         }
 
-        if (null === $passwordEncoder) {
-            $passwordEncoder = new PlaintextPasswordEncoder();
+        if (null === $passwordHasher) {
+            $passwordHasher = new PlaintextPasswordHasher();
         }
 
-        $encoderFactory = $this->createMock(EncoderFactoryInterface::class);
-        $encoderFactory
+        $hasherFactory = $this->createMock(PasswordHasherFactoryInterface::class);
+        $hasherFactory
             ->expects($this->any())
-            ->method('getEncoder')
-            ->willReturn($passwordEncoder)
+            ->method('getPasswordHasher')
+            ->willReturn($passwordHasher)
         ;
 
-        return new DaoAuthenticationProvider($userProvider, $userChecker, 'key', $encoderFactory);
+        return new DaoAuthenticationProvider($userProvider, $userChecker, 'key', $hasherFactory);
     }
 }