+++ src/Symfony/Component/Security/Http/EventListener/PasswordMigratingListener.php
@@ -12,12 +12,15 @@
 namespace Symfony\Component\Security\Http\EventListener;
 
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\PasswordHasher\Exception\InvalidPasswordException;
 use Symfony\Component\Security\Core\Encoder\EncoderFactoryInterface;
+use Symfony\Component\Security\Core\Exception\BadCredentialsException;
 use Symfony\Component\Security\Core\User\PasswordUpgraderInterface;
 use Symfony\Component\Security\Http\Authenticator\Passport\Badge\PasswordUpgradeBadge;
 use Symfony\Component\Security\Http\Authenticator\Passport\Badge\UserBadge;
 use Symfony\Component\Security\Http\Authenticator\Passport\UserPassportInterface;
 use Symfony\Component\Security\Http\Event\LoginSuccessEvent;
+use Symfony\Component\PasswordHasher\Hasher\PasswordHasherFactoryInterface;
 
 /**
  * @author Wouter de Jong <wouter@wouterj.nl>
@@ -27,11 +30,18 @@ use Symfony\Component\Security\Http\Event\LoginSuccessEvent;
  */
 class PasswordMigratingListener implements EventSubscriberInterface
 {
-    private $encoderFactory;
+    private $hasherFactory;
 
-    public function __construct(EncoderFactoryInterface $encoderFactory)
+    /**
+     * @param PasswordHasherFactoryInterface $hasherFactory
+     */
+    public function __construct($hasherFactory)
     {
-        $this->encoderFactory = $encoderFactory;
+        if ($hasherFactory instanceof EncoderFactoryInterface) {
+            trigger_deprecation('symfony/security-core', '5.3', 'Passing a "%s" instance to the "%s" constructor is deprecated, use "%s" instead.', EncoderFactoryInterface::class, __CLASS__, PasswordHasherFactoryInterface::class);
+        }
+
+        $this->hasherFactory = $hasherFactory;
     }
 
     public function onLoginSuccess(LoginSuccessEvent $event): void
@@ -50,8 +60,8 @@ class PasswordMigratingListener implements EventSubscriberInterface
         }
 
         $user = $passport->getUser();
-        $passwordEncoder = $this->encoderFactory->getEncoder($user);
-        if (!$passwordEncoder->needsRehash($user->getPassword())) {
+        $passwordHasher = $this->hasherFactory instanceof EncoderFactoryInterface ? $this->hasherFactory->getEncoder($user) : $this->hasherFactory->getPasswordHasher($user);
+        if (!$passwordHasher->needsRehash($user->getPassword())) {
             return;
         }
 
@@ -72,7 +82,7 @@ class PasswordMigratingListener implements EventSubscriberInterface
             }
         }
 
-        $passwordUpgrader->upgradePassword($user, $passwordEncoder->encodePassword($plaintextPassword, $user->getSalt()));
+        $passwordUpgrader->upgradePassword($user, $passwordHasher->hash($plaintextPassword, $user->getSalt()));
     }
 
     public static function getSubscribedEvents(): array